pyproject.toml
[tool.poetry]
name = "tempren"
version = "0.9.0"
description = "Template-based renaming utility"
readme = "README.md"
license = "GPL-3.0-or-later"
authors = ["Paweł Żukowski <p.z.idlecode@gmail.com>"]
homepage = "https://github.com/idle-code/tempren"
keywords = ["batch-renaming", "cli", "filename"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "Natural Language :: English",
    "Operating System :: POSIX",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "^4.10"
Unidecode = "^1.2.0"
pathvalidate = "^2.4.1"
mutagen = "^1.45.1"
docstring-parser = "^0.13"
python-magic = "^0.4.27"
Pillow = "^9.2.0"
piexif = "^1.1.3"
pymediainfo = {version = "^5.1.0", optional = true}
isodate = "^0.6.1"
gpxpy = "^1.5.0"
Pint = "^0.20.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1"
pytest-cov = "^2.10.1"
coverage = "^5.3"
pre-commit = "^2.7.1"
mypy = "^0.790"

[tool.poetry.extras]
video = ["pymediainfo"]

[tool.poetry.scripts]
tempren = "tempren.cli:throwing_main"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pytest.ini_options]
addopts = "--cov-report=xml --cov-report=term-missing --cov=tempren"
testpaths = ["tests"]
markers = [
    "e2e: slower, end-to-end tests often running tempren as separate process",
    "slow: slower test that often use timeouts"
]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.masonry.api"
